+++ src/main/java/com/wfm/experts/setup/wfm/leavepolicy/dto/RepeatedlyGrantDetailsDto.java
@@ -1,3 +1,4 @@
+// src/main/java/com/wfm/experts/setup/wfm/leavepolicy/dto/RepeatedlyGrantDetailsDto.java
 package com.wfm.experts.setup.wfm.leavepolicy.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -14,9 +15,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class RepeatedlyGrantDetailsDto {
-    private Integer maxDaysPerYear;
-    private Integer maxDaysPerMonth;
-    private Integer maxDaysPerPayPeriod; // Added field
+    private Double maxDaysPerYear;
+    private Double maxDaysPerMonth;
+    private Double maxDaysPerPayPeriod; // Added field
     private GrantPeriod grantPeriod; // Added field
     private PostingType posting; // Added field
     private Integer minAdvanceNoticeInDays;
